Split assertion file into separate files
