+++ src/assertions/toBeDefined.mjs
@@ -1,0 +1,10 @@
+import { AssertionError } from '../errors/assertion.mjs'
+import { indentLine } from '../transform.mjs'
+
+export const toBeDefined = (actual) => {
+  if (actual === undefined) {
+    throw new AssertionError(indentLine('Received: <actual>'), {
+      actual,
+    })
+  }
+}
