+++ src/assertions/toBeFalsy.mjs
@@ -1,0 +1,16 @@
+import { AssertionError } from '../errors/assertion.mjs'
+import { RunnerError } from '../errors/runner.mjs'
+import { indentLine } from '../transform.mjs'
+
+export const toBeFalsy = (actual) => {
+  if (typeof actual !== 'boolean') {
+    throw new RunnerError(
+      indentLine(`Received: ${actual} is not a boolean value`)
+    )
+  }
+  if (actual !== false) {
+    throw new AssertionError(indentLine('Received: <actual> to be false'), {
+      actual,
+    })
+  }
+}
