+++ src/assertions/toBeUndefined.mjs
@@ -1,0 +1,10 @@
+import { AssertionError } from '../errors/assertion.mjs'
+import { indentLine } from '../transform.mjs'
+
+export const toBeUndefined = (actual) => {
+  if (actual !== undefined) {
+    throw new AssertionError(indentLine('Recieved: <actual> to be undefined'), {
+      actual,
+    })
+  }
+}
