+++ src/expect.mjs
@@ -2,7 +2,7 @@
 /* eslint-disable indent */
 import { EOL } from 'os'
 import { AssertionError } from './errors/assertion.mjs'
-import * as assertions from './assertions.mjs'
+import * as assertions from './assertions/assertions.mjs'
 import { RunnerError } from './errors/runner.mjs'
 import { indentLine } from './transform.mjs'
